+++ src/umpire/strategy/DynamicPoolMap.hpp
@@ -71,6 +71,7 @@ class DynamicPoolMap : public AllocationStrategy,
   void release() override;
 
   std::size_t getActualSize() const noexcept override;
+  std::size_t getCurrentSize() const noexcept override;
 
   Platform getPlatform() noexcept override;
 
@@ -178,6 +179,7 @@ class DynamicPoolMap : public AllocationStrategy,
   const std::size_t m_next_minimum_pool_allocation_size;
 
   std::size_t m_actual_bytes{0};
+  std::size_t m_current_bytes{0};
   bool m_is_destructing{false};
 };